crypto: tcrypt: new user interface with debugfs
Tcrypto is a kernel module which can be used to trigger speed tests on
different parts of the crypto API. To trigger a test, users has to load the
module using various module parameters, e.g.:

 # modprobe tcrypt mode=404 sec=5

All tcrypt tests are implemented in the _init function, their output is sent to
the printk buffer, and once the test is over _init always returns a failure.

This patch modifies tcrypt in order to expose a debugfs user space interface
which can be used to trigger tcrypt tests and to get their output. All tests
output is redirected to an internal ring buffer.

The following files are available in the tcrypt debugfs directory:

 - command: this file can be use to control which test should be run simply
            writing the test name to it. it also supports clearing the 'output'
            buffer and listing all available tests

 - output:  holds a ring buffer with the output of the tests

 - seconds: this file can be used to control the number of seconds that a
            tests should be run consecutively. If zero it uses CPU cycles

This patch is not for inclusion, but I'd like to know if debugfs is the right
place to expose tcrypt interface, and if my approach makes sense.

Some functionalities available in the current version of the module are still
missing in this version (in particular test_available and do_alg_test are not
implemented yet). I also don't like the way in which I implemented the
'fixed_tests' but as i said this is just a prototype.

Signed-off-by: Luca Clementi <luca.clementi@gmail.com>